null check changes

+++ src/main/java/com/hcl/hawkeye/resourcemanagement/controller/ResourceManagementController.java
@@ -58,9 +58,9 @@ private static final Logger logger = LoggerFactory.getLogger(ResourceManagementC
 	
 	@RequestMapping(value="/getResourcesPercentByPortfolio/{portfolioId}", method = RequestMethod.GET, produces = "application/json")
 	@ResponseBody
-	public Integer getResourcesPercentByPortfolio(@PathVariable("portfolioId") int portfolioId) {
+	public Double getResourcesPercentByPortfolio(@PathVariable("portfolioId") int portfolioId) {
 		logger.info("Requested to get the resources count based on programId");
-		Integer response = resourceManagementService.getResourcesPercentByPortfolio(portfolioId);
+		Double response = resourceManagementService.getResourcesPercentByPortfolio(portfolioId);
 		return response;
 	}
 }
